feat: add GPU support for MSM
When enabling the `cuda` or `opencl` feature, then the GPU is used
to run the MSM. Currently there's no automatic fallback to CPU (this
is the case so during testing, we don't accidentally run on the CPU).

On CI, don't run with `--all-features` as the CI doesn't have GPU
support anyway.